+++ tests/unit/services/ComponentsTest.php
@@ -139,6 +139,13 @@ class ComponentsTest extends Unit
         $this->assertStringContainsString('data-hx-vals="{&quot;xYZ&quot;:&quot;a&quot;,&quot;limit&quot;:1}"', $html);
     }
 
+    public function testGetParsedTagAttributesValsWithBrackets()
+    {
+        $html = '<div sprig s-val:fields[x-y-z]="a"></div>';
+        $html = Sprig::$core->components->parse($html);
+        $this->assertStringContainsString('data-hx-vals="{&quot;fields[xYZ]&quot;:&quot;a&quot;}"', $html);
+    }
+
     public function testGetParsedTagAttributesValsEncodedAndSanitized()
     {
         $html = '<div sprig s-val:x="alert(\'xss\')" s-val:z=\'alert("xss")\' s-vals=\'{"limit":1}\'></div>';